This branch contains code that uses: Bundle bundle = data.getExtras(); Bitmap bitmap = (Bitmap) bundle.get("data");
as opposed to:
InputStream is = getContentResolver().openInputStream(data.getData());
Bitmap bitmap = BitmapFactory.decodeStream(is);

when retireving the informaiton from the camera intent in the onActivity Result method. This strategy works when running on my phone where (previously data.getData would only return null) however the image returned is a pixelated version of the actual image.